Refactor helpers into util.js
