+++ src/webgpu/shader/validation/decl/var_access_mode.spec.ts
@@ -16,9 +16,9 @@ import {
 } from '../../types.js';
 import { ShaderValidationTest } from '../shader_validation_test.js';
 
-import { declareEntryPoint, ShaderStage } from './util.js';
+import { declareEntryPoint, getVarDeclShader, accessModeExpander, infoExpander, supportsRead, supportsWrite, ShaderStage } from './util.js';
 
-// Test address spaces that we can hold an i32 variable.
+// Address spaces that can hold an i32 variable.
 const kNonHandleAddressSpaces = keysOf(kAddressSpaceInfo).filter(
   as => as !== 'handle'
 ) as AddressSpace[];
@@ -26,74 +26,11 @@ const kNonHandleAddressSpaces = keysOf(kAddressSpaceInfo).filter(
 export const g = makeTestGroup(ShaderValidationTest);
 
 /**
- * @returns a WGSL var declaration with given parameters for variable 'x' and
- * store type i32.
- */
-function declareVarX(
-  addressSpace: AddressSpace | undefined,
-  accessMode: AccessMode | undefined
-): string {
-  const parts: string[] = [];
-  if (addressSpace && kAddressSpaceInfo[addressSpace].binding) parts.push('@group(0) @binding(0) ');
-  parts.push('var');
-
-  const template_parts: string[] = [];
-  if (addressSpace) template_parts.push(addressSpace as string);
-  if (accessMode) template_parts.push(accessMode);
-  if (template_parts.length > 0) parts.push(`<${template_parts.join(',')}>`);
-
-  parts.push(' x: i32;');
-  return parts.join('');
-}
-
-/**
- * @returns a WGSL program for the test parameterization.
- */
-function getShader(
-  p: {
-    addressSpace: AddressSpace; // What address space for the variable.
-    explicitSpace: boolean; // Should the address space be explicitly spelled?
-    accessMode?: AccessMode; // What access mode to use.
-    explicitMode: boolean; // Should the access mode be explicitly spelled?
-    stage: ShaderStage; // What shader stage to use.
-  },
-  additionalBody?: string
-): string {
-  const as_info = kAddressSpaceInfo[p.addressSpace];
-  const decl = declareVarX(
-    p.explicitSpace ? p.addressSpace : undefined,
-    p.explicitMode ? p.accessMode : undefined
-  );
-
-  additionalBody = additionalBody ?? '';
-
-  switch (as_info.scope) {
-    case 'module':
-      return decl + '\n' + declareEntryPoint({ stage: p.stage, body: additionalBody });
-
-    case 'function':
-      return declareEntryPoint({ stage: p.stage, body: decl + '\n' + additionalBody });
-  }
-}
-
-/** @returns the list of address space info objects for the given address space.  */
-function infoExpander(p: { addressSpace: AddressSpace }): readonly AddressSpaceInfo[] {
-  return [kAddressSpaceInfo[p.addressSpace]];
-}
-
-/**
  * @returns a list of booleans indicating valid cases of specifying the address
  * space.
  */
 function explicitSpaceExpander(p: { info: AddressSpaceInfo }): readonly boolean[] {
   return p.info.spell === 'must' ? [true] : [true, false];
-}
-/** @returns a list of usable access modes under given experiment conditions.  */
-function accessModeExpander(p: {
-  explicitMode: boolean; // Whether the access mode will be emitted.
-  info: AddressSpaceInfo;
-}): readonly AccessMode[] {
-  return p.explicitMode && p.info.spellAccessMode !== 'never' ? p.info.accessModes : [];
 }
 
 /**
@@ -113,21 +50,6 @@ export function varDeclCompatibleAddressSpace(p: {
   return !(p.info.spell === 'must') || p.explicitSpace;
 }
 
-/** @returns the effective access mode for the given experiment.  */
-function effectiveMode(p: { info: AddressSpaceInfo; accessMode: AccessMode }): AccessMode {
-  return p.accessMode || p.info.accessModes[0]; // default is first.
-}
-/** @returns whether the setup allows reads */
-function supportsRead(p: { info: AddressSpaceInfo; accessMode: AccessMode }): boolean {
-  const mode = effectiveMode(p);
-  return p.info.accessModes.includes(mode) && kAccessModeInfo[mode].read;
-}
-/** @returns whether the setup allows writes */
-function supportsWrite(p: { info: AddressSpaceInfo; accessMode: AccessMode }): boolean {
-  const mode = effectiveMode(p);
-  return p.info.accessModes.includes(mode) && kAccessModeInfo[mode].write;
-}
-
 g.test('explicit_access_mode')
   .desc('Validate uses of an explicit access mode on a var declaration')
   .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
@@ -143,7 +65,7 @@ g.test('explicit_access_mode')
         .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
   )
   .fn(t => {
-    const prog = getShader(t.params);
+    const prog = getVarDeclShader(t.params);
 
     const ok =
       // The address space must be explicitly specified.
@@ -169,7 +91,7 @@ g.test('implicit_access_mode')
         .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
   )
   .fn(t => {
-    const prog = getShader(t.params);
+    const prog = getVarDeclShader(t.params);
 
     // 7.3 var Declarations
     // "The access mode always has a default value,.."
@@ -192,7 +114,7 @@ g.test('read_access')
         .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
   )
   .fn(t => {
-    const prog = getShader(t.params, 'let copy = x;');
+    const prog = getVarDeclShader(t.params, 'let copy = x;');
     const ok = supportsRead(t.params);
     t.expectCompileResult(ok, prog);
   });
@@ -211,7 +133,7 @@ g.test('write_access')
         .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
   )
   .fn(t => {
-    const prog = getShader(t.params, 'x = 0;');
+    const prog = getVarDeclShader(t.params, 'x = 0;');
     const ok = supportsWrite(t.params);
     t.expectCompileResult(ok, prog);
   });
